requirements.txt
sgiref~=3.7.2
click~=8.1.3
Django~=4.2.2
django-cors-headers~=4.1.0
djangorestframework~=3.14.0
joblib~=1.2.0
nltk~=3.8.1
numpy~=1.25.0
pip~=23.1.2
pytz~=2023.3
regex~=2023.6.3
scikit-learn~=1.2.2
scipy~=1.11.0
setuptools~=67.8.0
sqlparse~=0.4.4
threadpoolctl~=3.1.0
tqdm~=4.65.0
wheel~=0.40.0